+++ BlazorDeviceControl/Shared/Item/Printer.razor.cs
@@ -1,9 +1,9 @@
-﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
-// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+﻿// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
 using BlazorCore;
 using BlazorCore.DAL;
 using BlazorCore.DAL.TableModels;
+using BlazorCore.Models;
 using BlazorCore.Utils;
 using Microsoft.AspNetCore.Components;
 using System.Collections.Generic;
@@ -16,7 +16,7 @@ namespace BlazorDeviceControl.Shared.Item
     {
         #region Public and private fields and properties
 
-        private PrinterEntity PrinterItem => IdItem is PrinterEntity idItem ? idItem : null;
+        public PrinterEntity PrinterItem { get => (PrinterEntity)IdItem; set => SetItem(value); }
         public List<PrinterTypeEntity> PrinterTypeItems { get; set; } = null;
 
         #endregion
@@ -29,11 +29,12 @@ namespace BlazorDeviceControl.Shared.Item
             RunTasks($"{LocalizationStrings.DeviceControl.Method} {nameof(SetParametersAsync)}", "", LocalizationStrings.Share.DialogResultFail, "",
                 new List<Task> {
                     new(async() => {
-                        IdItem = null;
+                        Table = new TableScaleEntity(EnumTableScale.Printers);
+                        PrinterItem = null;
                         PrinterTypeItems = null;
                         await GuiRefreshWithWaitAsync();
 
-                        IdItem = AppSettings.DataAccess.PrintersCrud.GetEntity(new FieldListEntity(new Dictionary<string, object>
+                        PrinterItem = AppSettings.DataAccess.PrintersCrud.GetEntity(new FieldListEntity(new Dictionary<string, object>
                             { { EnumField.Id.ToString(), Id } }), null);
                         PrinterTypeItems = AppSettings.DataAccess.PrinterTypesCrud.GetEntities(null, null).ToList();
                         await GuiRefreshWithWaitAsync();