+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.450] - 2021-08-31
+### Changed
+- BlazorCore v.0.5.450
+- BlazorDeviceControl v.0.5.450
+
 ## [0.5.440] - 2021-08-23
 ### Changed
 - SsccWebAPI v.0.5.440